% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smtools.R
\name{convert.dates}
\alias{convert.dates}
\title{Convert all columns with "date" in col names to Date format.}
\usage{
convert.dates(dt, additional.cols = NULL, format = "\%Y-\%m-\%d")
}
\arguments{
\item{dt}{Data table.}

\item{additional.cols}{Additional date columns to be parsed other than columns 
with the string "date".}
}
\description{
Convert all columns with "date" in col names to Date format.
}
\details{
\code{convert.dates} automatically finds columns with "date" in the 
name and converts to \code{Date}.
Additional columns can be parsed on for parsing. Cannot handle multiple date 
formats.
}
