% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smtools.R
\name{cp}
\alias{cp}
\title{Copy to clipboard}
\usage{
cp(x, row.names = FALSE, excel.names = TRUE)
}
\arguments{
\item{x}{Data to copy to clipboard.}

\item{row.names}{Whether to include row names. Default to \code{FALSE}.}

\item{excel.names}{Whether to convert col names to Title Case. Default to 
\code{TRUE}.}
}
\description{
Copy to clipboard
}
