% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smtools.R
\name{excel.names}
\alias{excel.names}
\title{Convert col names to title names}
\usage{
excel.names(dt)
}
\arguments{
\item{x}{Data to convert col names back to Proper Title Case.}
}
\description{
Convert col names to title names
}
