% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smtools.R
\name{fuzzymatch}
\alias{fuzzymatch}
\title{Fuzzy match; suggest possible match results}
\usage{
fuzzymatch(x, y, matchdt = NULL, n = 2, ...)
}
\arguments{
\item{x}{A vector to be matched.}

\item{y}{A vector to find the match from.}

\item{matchdt}{Optional. A data table with the first column being \code{y}.
This is to help facilitate fuzzy match to raw names of alread matched names,
and then merge to master names in one go.}

\item{n}{Number of match results; default to 2.}

\item{...}{Further arguments to be passed onto \code{stringdistmatrix}.}
}
\description{
Fuzzy match; suggest possible match results
}
\details{
No debugging / error catching. Used \code{stringdistmatrix} from 
\code{stringdist}.
}
