% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smtools.R
\name{gen.headers}
\alias{gen.headers}
\title{Generate sript headers.}
\usage{
gen.headers(title = "WRITE STH NICE")
}
\arguments{
\item{title}{Will convert to upper letters automatically.}
}
\description{
Generate sript headers.
}
