% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smtools.R
\name{\%Like\%}
\alias{\%Like\%}
\title{Case-insensitive %like%}
\usage{
x \%Like\% pattern
}
\arguments{
\item{x}{String to be detected.}

\item{pattern}{String pattern (regex), case insensitive.}
}
\description{
Case-insensitive %like%
}
