% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smtools.R
\name{ifunique}
\alias{ifunique}
\title{Examine whether one column in a data table is unique by another}
\usage{
ifunique(x, ...)
}
\arguments{
\item{x}{A list of column names or one column name without quote; the key
which other columns should be unique by.}

\item{...}{Other columns to investigate uniqueness.}
}
\description{
Examine whether one column in a data table is unique by another
}
\details{
No debugging / error catching.
}
