% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smtools.R
\name{months.diff}
\alias{months.diff}
\title{Difference of dates in months}
\usage{
\method{months}{diff}(date, refdate, since = TRUE)
}
\arguments{
\item{refdate}{Reference date in the same format.}

\item{since}{If \code{TRUE}, months since date to reference date, eg from 
date to now (ref), else
month from reference date to date, eg from now (ref) to a future dat)}
}
\description{
Difference of dates in months
}
