% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smtools.R
\name{parse2v}
\alias{parse2v}
\title{Save some time typing ". Parse one string to a vector.}
\usage{
parse2v(string, sep = ",")
}
\arguments{
\item{string}{A string to be parsed.}

\item{sep}{Default \code{,}.}
}
\description{
Save some time typing ". Parse one string to a vector.
}
