% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smtools.R
\name{pgroup}
\alias{pgroup}
\title{Pivot table inspired grouping.}
\usage{
pgroup(col, strgroup, suffix = ".grp", new = FALSE)
}
\arguments{
\item{col}{Vector; use in \code{data.table}.}

\item{strgroup}{Grouping string in the format of \code{"Other+Unknown=Other"}.
Supports fuzzy match. \code{"O+U=Other"}, if Other and Unknown are the only 
categories starting with O and U. Supports \code{"NA+U=Other"}. For blanks, 
use \code{"$+O=Other"}.}

\item{suffix}{Suffix of created grouped column.}

\item{new}{Create a new column of the grouped categories. Default to 
\code{FALSE}.}
}
\description{
Pivot table inspired grouping.
}
\details{
No debugging / error catching.
}
