% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smtools.R
\name{set.diff}
\alias{set.diff}
\title{Set difference between two vectors; imagine the Venn Diagram.}
\usage{
set.diff(x, y, res = -1)
}
\arguments{
\item{x}{A vector.}

\item{y}{A vector.}

\item{res}{Return data; see \code{value}.}
}
\description{
Set difference between two vectors; imagine the Venn Diagram.
}
