% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smtools.R
\name{str.pattern}
\alias{str.pattern}
\title{Detect string pattern.}
\usage{
\method{str}{pattern}(col, pattern = "Aa0")
}
\arguments{
\item{col}{A vector of string.}

\item{pattern}{Default to \code{Az0}. Supports \code{Aa0.} where each 
transforms all big letters,
all small letters, all numbers, everything else to \code{A, a, 0} 
and \code{.}.}
}
\description{
Detect string pattern.
}
